database push
database with reader of bmp header.
"5.bmp" is in reality a png file "5.png". It's just to show what it does if a .bmp file isn't a bitmap file with proper header.
There is also a two files number 9: 9.bmp and 9.jpeg to show it can handle both of them.